requirements.txt
tensorflow==2.1.0
tensorflow-gpu==2.1.0
pyqt==5.9.2
keras==2.3.1
keras-gpu==2.3.1
numpy>=1.19.2
scipy>=1.5.2
pandas>=1.1.3